i forgor

+++ src/main/java/net/ccbluex/liquidbounce/injection/forge/mixins/entity/MixinEntityLivingBase.java
@@ -5,6 +5,7 @@
  */
 package net.ccbluex.liquidbounce.injection.forge.mixins.entity;
 
+import de.enzaxd.viaforge.ViaForge;
 import net.ccbluex.liquidbounce.LiquidBounce;
 import net.ccbluex.liquidbounce.event.JumpEvent;
 import net.ccbluex.liquidbounce.features.module.modules.movement.AirJump;
@@ -119,7 +120,7 @@ public abstract class MixinEntityLivingBase extends MixinEntity {
     }
 
     @ModifyConstant(method = "onLivingUpdate", constant = @Constant(doubleValue = 0.005D))
-    private int refactor1_9MovementThreshold(int constant) {
+    private double refactor1_9MovementThreshold(int constant) {
         if (ViaForge.getInstance().getVersion() != 47)
             return 0.003D;
         return 0.005D;
